Whitelist TrieMap from explicit nulls
TrieMap is a tricky one to migrate: it's concurrent, it's a (complicated) trie map,
it interacts with Java, and as common for collections it uses `null` internally
for efficiency. With explicit nulls, there are many type errors and we've had
trouble understanding which fields are nullable and which aren't.

Instead of guessing (likely incorrectly) what should be nullable and
what shouldn't, "whitelist" the file by importing

scala.language.implicitNulls

This adds some "implicit" conversions involving null. With the
conversions, we almost don't have to change anything in the file
for it type.

So really this is a test case for the implicitNulls import.